Add : Connect View Controller
