--- src/TestSmartDeviceAPPs/iOS/IoTDeviceController/IoTDeviceController/ServerConnectViewController.swift
@@ -1,35 +1,0 @@
-//
-//  ServerConnectViewController.swift
-//  IoTDeviceController
-//
-//  Created by Kim, Min Ho on 2016. 2. 17..
-//  Copyright © 2016년 Luna. All rights reserved.
-//
-
-import UIKit
-
-class ServerConnectViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
-        // Get the new view controller using segue.destinationViewController.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
